refactor: reviewModal 위치 수정
- RatePlanStorePage -> ChatbotComponent에 포함되게 수정
